+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 * Android WebView file inputs use `file_picker`; capture requests use
   `image_picker` (camera).
 * Queue `postHostEvent` until the widget page has finished loading.
-* Export configuration / events from `package:ramp_flutter/ramp_flutter.dart`.
+* Public API is `package:ramp_flutter/ramp_flutter.dart` only (implementation under `lib/src/`).
 
 ## 4.0.1
