+++ lib/src/configuration.dart
