+++ lib/src/host_event.dart
