+++ lib/src/models/purchase_details.dart
@@ -1,5 +1,5 @@
-import 'package:ramp_flutter/models/asset_info.dart';
-import 'package:ramp_flutter/models/json_map.dart';
+import 'package:ramp_flutter/src/models/asset_info.dart';
+import 'package:ramp_flutter/src/models/json_map.dart';
 
 class PurchaseDetails {
   const PurchaseDetails({
