+++ lib/src/ramp_webview.dart
@@ -4,8 +4,8 @@ import 'dart:convert';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/widgets.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:ramp_flutter/host_event.dart';
-import 'package:ramp_flutter/widget_event.dart';
+import 'package:ramp_flutter/src/host_event.dart';
+import 'package:ramp_flutter/src/widget_event.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 import 'package:webview_flutter_android/webview_flutter_android.dart';
