+++ lib/src/widget_events/widget_close.dart
