+++ lib/src/widget_events/widget_close_request.dart
